+++ src/resource/createResource.composable.ts
@@ -1,3 +1,4 @@
+import type { ResourceType } from "@/api/api.types";
 import { useAuth } from "@/auth/auth.composable";
 import { useResourceStore } from "@/store/resource.store";
 
@@ -5,15 +6,19 @@ export default function useCreateResource() {
   const { refreshJwt } = useAuth();
   const resourceStore = useResourceStore();
 
-  async function addNewResource(resourceId: string) {
+  /**
+   * Register a newly created resource.
+   * @param type The type is necessary for the store's filter getters to include the new object.
+   * @param resourceId The resource id.
+   */
+  async function addNewResource(type: ResourceType, resourceId: string) {
     // Have the new corpus included in further API calls.
     await refreshJwt();
 
     // Adding the new id to store may trigger API calls, so do it after updating the JWT.
-    resourceStore.resources[resourceId] =
-      resourceStore.resources[resourceId] || {};
-
-    return resourceId;
+    if (!(resourceId in resourceStore.resources)) {
+      resourceStore.resources[resourceId] = { type };
+    }
   }
 
   return {